feat(app): Add "About 42Summaries" menu item and window
This change adds a new "About 42Summaries" menu item to the application's main menu, and implements a corresponding `AboutView` window that displays information about the app. The changes include:

- Adding the `AppDelegate` class to handle the setup of the "About" menu item and window
- Implementing the `showAboutView()` function in `AppDelegate` to display the "About" window when the menu item is selected
- Creating the `AboutView` SwiftUI view to display the app's information
- Updating the `Info.plist` file to include user notification settings

The motivation for these changes is to provide users with easy access to information about the 42Summaries app, including its version, copyright, and other relevant details.