Added AIModelRetriever to project and refactored model retrieval
The new AIModelRetriever package has been added to the project. This package is used for retrieving models from different providers. The SettingsView has been refactored to use this new package, simplifying the code by abstracting away the details of each provider's API. The available models are now retrieved using the AIModelRetriever, which provides a unified interface for fetching models regardless of the provider.